requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
prophet==1.1.5

# Data Processing
python-dateutil==2.8.2
pytz==2023.3

# Database
sqlalchemy==2.0.23
alembic==1.12.1
sqlite3

# Dashboard
streamlit==1.28.1
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# API Utilities
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging
loguru==0.7.2

# Utilities
tqdm==4.66.1
pydantic-extra-types==2.3.0

